+++ src/periodic.h
@@ -12,7 +12,7 @@ namespace periodic {
 
 	std::string polyalphabeticDecrypt(std::string cipher, std::vector<std::array<char, 26>>, bool);
 
-	std::string hillClimber(std::string cipher, int keySize);
+	std::string hillClimber(std::string cipher, long keySize);
 	std::string hillClimber(std::string cipher);
 
 	int cliHillClimber(std::string cipher);
